fill tg chat adapter

+++ core/src/main/java/io/github/_127_0_0_l/core/models/TgChat.java
@@ -1,7 +1,7 @@
 package io.github._127_0_0_l.core.models;
 
 public record TgChat(
-        int chatId,
+        Long chatId,
         boolean isActive,
         ChatState state,
         Filters filters
